restructure model code
