+++ docs/pix2tex.models.rst
@@ -1,7 +1,15 @@
-pix2tex.structures.hybrid package
-======================
+pix2tex.models.hybrid package
+=============================
 
-.. automodule:: pix2tex.structures.hybrid
+.. automodule:: pix2tex.models.hybrid
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+pix2tex.models.vit package
+==========================
+
+.. automodule:: pix2tex.models.vit
    :members:
    :no-undoc-members:
    :show-inheritance:
